Made `Switch`  menu item smarter.
From my experience, I had some situations where I though I minimized
WebCord to the tray and tried to unminimize it,
but instead it were just burried under some windows.
This update should change that so clicking on `Switch` in tray
will also focus the app in case it is visible and it's not focused.