Implement permission management.
Previously, WebCord accepted automatically the required permissions
by Discord. As this wasn't the best sollution for the user privacy (i.e.
WebCord didn't really checked the reason of the permisssion request),
I've implemented a permission management via the settings window, with
some defaults to block most of the requests that could affect the
user privacy.