Better dektop capturer
Reimplemented the desktop capturer using "experimental" BrowserView API
to both prepare the app for Electron's 17 breaking changes and make it
to not break Discord's ToS anymore (since it is no longer injected on
the website). It is also much more secure, since Discord is unable to
read desktop capturer information about the sources anymore.