Improve API key generation.
I've decided to work again on the property generation to make it look
less suspicious and harder to find – previously, there was a great
chance to find it by property length, it's type and/or charset it uses.
I've also add an additional check to make sure the generated key is not
already the existing property name.