Suppress the update notifications on "close".
This commit adds some logic on suppressing the updates, so they will be
less annoying on more frequent updates. It will suppress update
notifications for a week if there's not going to be any newer version
released and user has manually closed the notification (on Linux,
it is done by clicking on it and not clicking on any of the buttons).