Types hardening.
This update switches to use the official TypeScript
`@tsconfig/node16-strictest` config as its base to make TSC "strict by
default". It also changes some Electron imports to indicate the process
in which they're meant to be used. Because of that, It obviously fixes some
types as well and updates `eslintrc`.