Try to mitigate the unsafe flags on production.
This change is meant to mitigate the use of *unsafe* flags from
`--inspect` family by removing them at runtime for production builds.
However, this is not meant to replace the *fuses* neither is tested
if it is functional, It should work fine as another layer of protection,
being "better than nothing", especially for unpatched Electron builds.