Fix performance issues with `CSP`.
The problem with recalculating the CSP every time is that WebCord
actually resend its headers for every element (scripts, images etc.) it
loads and for every request it mades. And generating the CSP takes quite
a time, mainly due to the loops. Therefore, I've previously added
a cache that I thought is should fix the problem.

However, this was not enough. While it mitigated the main issue, a lot
of unnecesary recalculations was still in the code. Therefore I've both
improved the CSP class implementation to calculate its value types at
initialization and use them in `toString` and `toObject` methods. I've
also improved the cache itself by ading a "helper" value to make it
possible to determine if cache is out-of-date even before
`getWebcordCSP` will enter any loops and try to calculate anything. This
way the code is kept performant on multiple requests while the result is
kept up-to-date with the user settings.