Initial "dynamic configuration" feture.
This commit simplifies the configuration structure to use a proper
AppConfig stucture and L10N to generate a GUI for configuration in
WebCord. Everything is done dynamically now and settings will always
update accordingly to the default config and default localization files.

Expected to break, since code is still not adapted to new configuration
structure.