Prepare WebCord for incoming Electron changes.
Currently, Electron either depends on Node in preloads or does use
CommonJS to even load the unpriviledged modules. This is why it has to
be disabled in most windows. Also please note `sandbox` is disabled by
the default and does not affect that `contextIsolation` still isolates
preloads from web content and normally there should be no way for
browser scripts to tamper with the preloads and access Node.

In the future, I may decide to enable sandboxing for main window only.
There's not much to adapt to make it functional, mostly it fails now as
`tslib` can't be loaded without Node and I need to resolve how to
"replace" CommonJS modules.