Initial implementation of themes encryption.
The encryption prevents third-party software to actually inject their
own themes (and in the future – replace a CSP that is needed for a
particular theme to work). While
I don't guarantee any app
couldn't bypass this, it
definitely makes it harder or
impossible depending if OS can prevent manipulation if displayed dialog
windows by another software.

This change is breaking for those
who used previously an unencrypted CSS themes; you will have to merge
your themes and let WebCord to either encrypt them or not based if
encryption is supported on given platform.