Tray: use more "stable" image format.
The `toBitmap` method is actually quite "unstable": it's behavior varies
from platform and produce different results. It's not reliable for hash
verifications, therefore a better format of an image should be used.

This commit changes the format to PNG and regenrates checksums for it.
This should produce much more reliable results than it was before.