Add keybind socket.
Because Wayland forbids processes from listening to system-wide
keypresses, Discord and WebCord are unable to listen for keybinds
when the window doesn't have focus. This commit adds a socket server
that allows for sending a payload to forward keybinds to WebCord. An
example of activating the push-to-talk key is:

`echo -e 'PUSH_TO_TALK\x00\x01\x99\x00' | nc -U $XDG_RUNTIME_DIR/webcord-keybinds.sock`

The payload begins with one of the keybind actions that Discord
supports, which is terminated with a null (\x00) byte. The byte that
follows is a boolean value - a true value (\x01) indicates a keyDown
event, and a false value (\x00) indicates a keyUp event. Note that
keybinds that aren't toggle-based, like PUSH_TO_TALK, generally only
respond to keyUp events.