WSS: Add some placeholder code to handle HTTP.
In Discord, WebSocket can reply to some HTTP requests done by the
browser. This code is meant to be a placeholder, with very basic REST
API implementation. I will work on it again once I'll analyze
the requests and responses made by Discord and try to reply to them in
some logical way.