Make `socket.ts` independent on main window.
This commit adds logic for `socket.ts` to find main windon on its own
and handle connections without it (i.e. reject all until main window is
present again).

This is one step done to achieve proper macOS support, with a
possibility to close window and keep app running in the background.